Module google_workflowexecutions1::api

Structs

CancelExecutionRequest
Request for the CancelExecution method.

Error
Error describes why the execution was abnormally terminated.

Execution
A running instance of a Workflow.

ListExecutionsResponse
Response for the ListExecutions method.

Position
Position contains source position information about the stack trace element such as line number, column number and length of the code block in bytes.

ProjectLocationWorkflowExecutionCancelCall
Cancels an execution of the given name.

ProjectLocationWorkflowExecutionCreateCall
Creates a new execution using the latest revision of the given workflow.

ProjectLocationWorkflowExecutionGetCall
Returns an execution of the given name.

ProjectLocationWorkflowExecutionListCall
Returns a list of executions which belong to the workflow with the given name. The method returns executions of all workflow revisions. Returned executions are ordered by their start time (newest first).

ProjectMethods
A builder providing access to all methods supported on project resources. It is not used directly, but through the WorkflowExecutions hub.

StackTrace
A collection of stack elements (frames) where an error occurred.

StackTraceElement
A single stack element (frame) where an error occurred.

WorkflowExecutions
Central instance to access all WorkflowExecutions related resource activities

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.